Add files via upload
This is a Python code for a Hangman game menu initializer. It includes a loop that allows the user to choose from three options: animals, countries or close the game. Once the user selects a category, the program will generate a hidden word from that category and display the user's progress in guessing the letters of the word. 

The user has 10 attempts to guess the word, and each incorrect guess will reduce the number of remaining attempts. The program will display the correct and incorrect letters that the user has guessed so far, as well as a visual representation of the Hangman game board. 

If the user guesses all the unique letters of the hidden word before running out of attempts, the program will congratulate the user. If the user runs out of attempts without guessing the word, the program will reveal the hidden word and inform the user that they have lost the game. 

After the user completes a game, the program will prompt the user to press ENTER to continue or start a new game by selecting a category.